lambdas/bucket_raw_data/requirements.txt
awswrangler
boto3
python-dotenv